% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adaboost_5_all}
\alias{adaboost_5_all}
\title{All atoms used model}
\format{
An object of class \code{train} (inherits from \code{train.formula}) of length 25.
}
\usage{
adaboost_5_all
}
\description{
Adaboost model built using 5Å cutoff and all atoms in the structure.
}
\keyword{datasets}
