% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amino_acid_features.R
\name{amino_acid_features}
\alias{amino_acid_features}
\title{Amino Acid Features}
\usage{
amino_acid_features(info_df)
}
\arguments{
\item{info_df}{data.frame contains all the input mutations}
}
\value{
accessible surface area and hydrophobicity scale
values of reference amino acid, mutant amino acid and difference between mutant and reference
amino acids
}
\description{
This function calculates accessible surface area and hydrophobicity scale
values of reference amino acid, mutant amino acid and difference between mutant and reference
amino acids.
}
