% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betweenness_score.R
\name{betweenness_score}
\alias{betweenness_score}
\title{Betweenness Score}
\usage{
betweenness_score(edge_list, filtered_info_df)
}
\arguments{
\item{edge_list}{list contains separate edge data.frames for each chain}

\item{filtered_info_df}{input data.frame which contain only one PDB entries}
}
\value{
Betweenness Z-Scores of the input position
}
\description{
Calculate Betweenness Z-Scores of input position
}
\details{
This function calculates Betweenness scores of all nodes in the network via \code{betweenness} function
of igraph package. Calculates and returns the Z-Scores of the input positions.
}
