% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{caret_adaboost_all}
\alias{caret_adaboost_all}
\title{All atoms used model}
\format{
An object of class \code{train} (inherits from \code{train.formula}) of length 25.
}
\usage{
caret_adaboost_all
}
\description{
Adaboost model built using 7Å cutoff and all atoms in the structure.
}
\keyword{datasets}
