% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clique_score.R
\name{clique_score}
\alias{clique_score}
\title{Clique Score}
\usage{
clique_score(edge_list, filtered_info_df, n_threads = NULL, single_run = TRUE)
}
\arguments{
\item{edge_list}{list contains separate edge data.frames for each chain}

\item{filtered_info_df}{input data.frame which contain only one PDB entries}

\item{n_threads}{number of threads (default = NULL)}

\item{single_run}{should be set as TRUE when using \code{clique_score} function alone (default = TRUE)}
}
\value{
Clique Z-Scores of input position
}
\description{
Calculates and returns the Clique Z-Scores of the input positions.
}
