% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisGeNET.R
\name{DisGeNET}
\alias{DisGeNET}
\title{Disease-Gene relationship from DisGeNET}
\usage{
DisGeNET(filtered_info_df)
}
\arguments{
\item{filtered_info_df}{input data.frame which contain only one PDB entries}
}
\value{
Number of diseases related with input gene
}
\description{
Function finds the number of diseases related with the input genes.
}
