% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_centrality_score.R
\name{eigen_centrality_score}
\alias{eigen_centrality_score}
\title{Eigen Centrality Score}
\usage{
eigen_centrality_score(connections_df, filtered_info_df)
}
\arguments{
\item{connections_df}{data.frame contains all the edges}

\item{filtered_info_df}{input data.frame which contain only one PDB entries}
}
\value{
Eigen Centrality Z-Score of input position
}
\description{
Calculate Eigen Centrality Z-Scores of input position
}
\details{
This function calculates Eigen Centrality scores of all nodes in the network.
Calculates and returns the Z-Scores of the input positions.
}
