% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_score.R
\name{eigen_score}
\alias{eigen_score}
\title{Eigen Centrality Score}
\usage{
eigen_score(connections_df, Position, atom_matrix)
}
\arguments{
\item{connections_df}{data.frame contains all the edges}

\item{Position}{Mutation position}

\item{atom_matrix}{matrix that contains all the atoms in the PDB file}
}
\value{
Eigen Centrality Z-Score of input position
}
\description{
Calculate Eigen Centrality Z-Scores of input position
}
\details{
This function calculates Eigen Centrality scores of all nodes in the network.
Calculates and returns the Z-Scores of the input positions.
}
