% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_essentiality.R
\name{gene_essentiality}
\alias{gene_essentiality}
\title{Gene Essentiality Score}
\usage{
gene_essentiality(filtered_info_df)
}
\arguments{
\item{filtered_info_df}{input data.frame which contain only one PDB entries}
}
\value{
Gene Essentiality Scores of input genes
}
\description{
Function finds the gene essentiality scores of the input genes from OGEE database.
}
