% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genic_intolerance.R
\name{genic_intolerance}
\alias{genic_intolerance}
\title{Genic Intolerance Score}
\usage{
genic_intolerance(gene_name)
}
\arguments{
\item{gene_name}{Gene name}
}
\value{
numeric
}
\description{
Genic Intolerance Score
}
