% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnomad_scores.R
\name{gnomad_scores}
\alias{gnomad_scores}
\title{GnomAD Scores}
\usage{
gnomad_scores(PDB_ID, info_df)
}
\arguments{
\item{PDB_ID}{PDB ID}

\item{info_df}{data.frame contains all the input mutations}
}
\value{
list that contains a data.frame containing GnomAD scores of input mutations and
the gene name
}
\description{
GnomAD Scores
}
