% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_terms.R
\name{GO_terms}
\alias{GO_terms}
\title{GO Term Number from Gene Name}
\usage{
GO_terms(filtered_info_df)
}
\arguments{
\item{filtered_info_df}{input data.frame which contain only one PDB entries}
}
\value{
Number of GO Terms which related with the input genes
}
\description{
Function finds the number of GO Terms which related with the input genes.
}
