% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_prediction.R
\name{imp_prediction}
\alias{imp_prediction}
\title{Impact Prediction}
\usage{
imp_prediction(final_df)
}
\arguments{
\item{final_df}{data.frame contains all the mutation information and the required information for impact prediction}
}
\value{
matrix
}
\description{
Impact Prediction
}
