% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_prediction.R
\name{impact_prediction}
\alias{impact_prediction}
\title{Impact Prediction}
\usage{
impact_prediction(final_df)
}
\arguments{
\item{final_df}{data.frame contains all the required features for impact prediction}
}
\value{
data.frame contains the prediction results
}
\description{
Function that makes the final prediction
}
\details{
This function takes a data.frame which contains all the required features and compute
the impact of a mutation.
}
