% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_type.R
\name{mutation_type}
\alias{mutation_type}
\title{Mutation Type Classifier}
\usage{
mutation_type(info_df)
}
\arguments{
\item{info_df}{data.frame contains all the input mutations}
}
\value{
Mutation Type as a factor; 1 = Deadly, 0 = Not deadly
}
\description{
This functios classifies the Charged-Charged, Charged-Polar, Polar-Charged, Polar-Polar and
Hydrophobic-Hydrophobic mutations as 'Not deadly (0)', classifies Charged-Hydrophobic,
Hydrophobic-Charged, Hydrophobic-Polar, Polar-Hydrophobic mutations as 'Deadly (1)'.
}
