% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDB_read.R
\name{PDB_read}
\alias{PDB_read}
\title{PDB Downloading and Reading}
\usage{
PDB_read(PDB_ID)
}
\arguments{
\item{PDB_ID}{PDB ID}
}
\value{
Matrix that contains all the atoms in the PDB structure
}
\description{
Download and read input PDB via \code{get.pdb} and \code{read.pdb} function of bio3d package
and filter the PDB file for only containing atoms.
}
