% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPI_network_scores.R
\name{PPI_network_scores}
\alias{PPI_network_scores}
\title{PPI Network Scores}
\usage{
PPI_network_scores(filtered_info_df)
}
\arguments{
\item{filtered_info_df}{input data.frame which contain only one PDB entries}
}
\value{
Network property z-scores of input gene name
}
\description{
PPI Network Scores
}
