% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predatoR.R
\name{predatoR}
\alias{predatoR}
\title{Mutation impact Prediction}
\usage{
predatoR(info_df, PDB_path = NULL, n_threads = NULL, gene_name_info = TRUE)
}
\arguments{
\item{info_df}{data.frame containing 'PDB_ID', 'Chain', 'Position',
'Reference Amino Acid', 'Mutated Amino Acid' and 'Gene_Name' (optional) information respectively.}

\item{PDB_path}{PDB file path (default = NULL)}

\item{n_threads}{number of threads (default = NULL)}

\item{gene_name_info}{whether there is gene name information in the input or not (default = TRUE)}
}
\value{
data.frame which contains prediction results
}
\description{
Main function for mutation impact prediction
}
\details{
With an input data.frame which contains 'PDB_ID', 'Chain', 'Position',
'Reference Amino Acid', 'Mutated Amino Acid' and 'Gene_Name' (optional) information respectively,
make a prediction about the mutation by using pre computed Adaboost model
and classifies the mutation as 'Disease Causing' or 'Silent'.
}
