% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSSM_Scores.R
\name{PSSM_Scores}
\alias{PSSM_Scores}
\title{PSSM Scores}
\usage{
PSSM_Scores(filtered_info_df)
}
\arguments{
\item{filtered_info_df}{input data.frame which contain only one PDB entries}
}
\value{
PSSM scores of input positions
}
\description{
PSSM Scores
}
