% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorteset_path_score.R
\name{shorteset_path_score}
\alias{shorteset_path_score}
\title{Shortest Path Score Function}
\usage{
shorteset_path_score(connections_df, Position)
}
\arguments{
\item{connections_df}{data frame contains list of edges}

\item{Position}{Mutation position}
}
\value{
integer
}
\description{
Calculates shortest path z-score of mutation index amino acid from list of edges
}
