% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betweenness_score.R
\name{betweenness_score}
\alias{betweenness_score}
\title{Betweenness Score}
\usage{
betweenness_score(connections_df, Position)
}
\arguments{
\item{connections_df}{data.frame contains all the edges}

\item{Position}{Mutation position}
}
\value{
numeric
}
\description{
Betweenness Score
}
