% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betweenness_score.R
\name{betweenness_score}
\alias{betweenness_score}
\title{Betweenness Score}
\usage{
betweenness_score(connections_df, Position)
}
\arguments{
\item{connections_df}{data.frame contains all the edges calculated by \code{PDB2connections} function}

\item{Position}{Mutation position}
}
\value{
Betweenness Z-Scores of the input position
}
\description{
Calculate Betweenness Z-Scores of input position
}
\details{
This function calculates Betweenness scores of all nodes in the network via \code{betweenness} function
of igraph package. Calculates and returns the Z-Scores of the input positions.
}
