% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLOSUM62_score.R
\name{BLOSUM62_score}
\alias{BLOSUM62_score}
\title{BLOSUM62 Score}
\usage{
BLOSUM62_score(info_df)
}
\arguments{
\item{info_df}{data.frame contains all the input mutations}
}
\value{
BLOSUM62 score of input mutation
}
\description{
Function returns BLOSUM62 scores of input mutations.
}
