% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blosum62_score.R
\name{blosum62_score}
\alias{blosum62_score}
\title{BLOSUM62 Score}
\usage{
blosum62_score(info_df)
}
\arguments{
\item{info_df}{data.frame contains all the input mutations}
}
\value{
numeric
}
\description{
BLOSUM62 Score
}
