% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree_score.R
\name{degree_score}
\alias{degree_score}
\title{Degree Score}
\usage{
degree_score(edge_list, filtered_info_df)
}
\arguments{
\item{edge_list}{list contains separate edge data.frames for each chain}

\item{filtered_info_df}{input data.frame which contain only one PDB entries}
}
\value{
Degree Z-Scores of input position
}
\description{
Degree Score is the total number of connections of the nodes. Calculates and returns the Z-Scores of the input positions.
}
