% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_centrality_score.R
\name{eigen_centrality_score}
\alias{eigen_centrality_score}
\title{Eigen Centrality Score}
\usage{
eigen_centrality_score(edge_list, filtered_info_df)
}
\arguments{
\item{edge_list}{list contains separate edge data.frames for each chain}

\item{filtered_info_df}{input data.frame which contain only one PDB entries}
}
\value{
Eigen Centrality Z-Score of input position
}
\description{
Calculate Eigen Centrality Z-Scores of input position
}
\details{
This function calculates the total number of connections of nodes to which a node is connected.
Calculates and returns the Z-Scores of the input positions.
}
