% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genic_intolerance.R
\name{genic_intolerance}
\alias{genic_intolerance}
\title{Genic Intolerance Score}
\usage{
genic_intolerance(filtered_info_df)
}
\arguments{
\item{filtered_info_df}{input data.frame which contain only one PDB entries}
}
\value{
Genic Intolerance scores added input data.frame
}
\description{
\code{genic_intolerance} function returns Genic Intolerance scores of input genes
}
