% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnomad.R
\name{gnomad}
\alias{gnomad}
\title{GnomAD Scores}
\usage{
gnomad(PDB_ID, info_df)
}
\arguments{
\item{PDB_ID}{PDB ID}

\item{info_df}{data.frame contains all the input mutations}
}
\value{
list
}
\description{
GnomAD Scores
}
