% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnomad_scores.R
\name{gnomad_scores}
\alias{gnomad_scores}
\title{GnomAD Scores}
\usage{
gnomad_scores(filtered_info_df)
}
\arguments{
\item{filtered_info_df}{input data.frame which contain only one PDB entries}
}
\value{
list that contains a data.frame containing GnomAD scores of input mutations and
the gene name
}
\description{
GnomAD Scores
}
