% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTEx.R
\name{GTEx}
\alias{GTEx}
\title{GTEx Median Expression Function}
\usage{
GTEx(filtered_info_df)
}
\arguments{
\item{filtered_info_df}{input data.frame which contain only one PDB entries}
}
\value{
median gene expression values of input genes
}
\description{
This function returns the input genes' median gene expression value of
54 different tissue type from \href{https://gtexportal.org/home/}{GTEx}.
}
