% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_prediction.R
\name{imp_prediction}
\alias{imp_prediction}
\title{Impact Prediction}
\usage{
imp_prediction(final_df)
}
\arguments{
\item{final_df}{data.frame contains all the mutation information and the required information for impact prediction}
}
\value{
matrix contains the prediction results
}
\description{
Function that makes the final prediction
}
\details{
This function takes a data.frame which contains all the required fields for impact prediction and compute
the prediction based on pre-computed Adaboost model.
}
