% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_prediction.R
\name{impact_prediction}
\alias{impact_prediction}
\title{Impact Prediction}
\usage{
impact_prediction(final_df, distance_cutoff = 7, network_approach = "all")
}
\arguments{
\item{final_df}{data.frame contains all the required features for impact prediction}

\item{distance_cutoff}{distance cutoff for setting edges (default = 7)}

\item{network_approach}{network building approach; "all" (default) for using all atoms or "ca" for using carbon alpha atoms only}
}
\value{
data.frame contains the prediction results
}
\description{
Function that makes the final prediction
}
\details{
This function takes a data.frame which contains all the required features and compute
the impact of a mutation.
}
