% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG_pathway_number.R
\name{KEGG_pathway_number}
\alias{KEGG_pathway_number}
\title{KEGG Pathway Number from Gene Name}
\usage{
KEGG_pathway_number(filtered_info_df)
}
\arguments{
\item{filtered_info_df}{input data.frame which contain only one PDB entries}
}
\value{
Number of KEGG pathways which contain the input genes
}
\description{
Function finds the number of KEGG Pathways which contains the input genes.
}
