% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg_pathway_number.R
\name{kegg_pathway_number}
\alias{kegg_pathway_number}
\title{KEGG Pathway Number from Gene Name}
\usage{
kegg_pathway_number(gene_name)
}
\arguments{
\item{gene_name}{Gene name}
}
\value{
Number of KEGG pathways which contain the input gene
}
\description{
KEGG Pathway Number from Gene Name
}
