% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_type.R
\name{mutation_type}
\alias{mutation_type}
\title{Mutation Type Classifier}
\usage{
mutation_type(info_df)
}
\arguments{
\item{info_df}{data.frame contains all the input mutations}
}
\value{
factor
}
\description{
Mutation Type Classifier
}
