% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDB2connections.R
\name{PDB2connections}
\alias{PDB2connections}
\title{Creating edge list from atom matrix}
\usage{
PDB2connections(atom_matrix)
}
\arguments{
\item{atom_matrix}{matrix that contains all the atoms in the PDB file}
}
\value{
data.frame
}
\description{
Creating edge list from atom matrix
}
