% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDB2connections.R
\name{PDB2connections}
\alias{PDB2connections}
\title{Creating edge list from atom matrix}
\usage{
PDB2connections(atom_matrix)
}
\arguments{
\item{atom_matrix}{matrix that contains all the atoms in the PDB file}
}
\value{
data.frame containing all the edges
}
\description{
Creates all the edges from the input atom matrix created by \code{PDB_read} function
}
\details{
This function calculates the distances between all the atoms in the atom matrix and create edges
if the distance between two atoms is less or equal to 7 Angstrom.
}
