% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredImption.R
\name{PredImption}
\alias{PredImption}
\title{Main mutation impact prediction function}
\usage{
PredImption(info_df)
}
\arguments{
\item{info_df}{data.frame contains 'PDB_ID', 'Chain', 'Position',
'Reference Amino Acid' and 'Mutated Amino Acid' information respectively}
}
\value{
data.frame
}
\description{
With an input data frame which contains 'PDB_ID', 'Chain', 'Position',
'Reference Amino Acid' and 'Mutated Amino Acid' information respectively,
make a prediction about the mutation with using adaboost model and classifies
the mutation as 'disease causing' or 'silent'.
}
