% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_PDB.R
\name{read_PDB}
\alias{read_PDB}
\title{PDB Reading/Downloading}
\usage{
read_PDB(PDB_ID, PDB_path = NULL, network_approach = "ca")
}
\arguments{
\item{PDB_ID}{PDB ID}

\item{PDB_path}{PDB file path (default = NULL)}

\item{network_approach}{network building approach; "ca" (default) for using ca atoms only or "all" for using all atoms}
}
\value{
Matrix that contains all the atoms in the PDB structure
}
\description{
Download/Read input PDB via \code{get.pdb} and \code{read.pdb} function of bio3d package
and filter the PDB file for only containing atoms.
}
