% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorteset_path_score.R
\name{shorteset_path_score}
\alias{shorteset_path_score}
\title{Shortest Path Score Function}
\usage{
shorteset_path_score(connections_df, Position)
}
\arguments{
\item{connections_df}{data frame contains list of edges}

\item{Position}{Mutation position}
}
\value{
Shortest Path Z-Scores of input position
}
\description{
Calculate Shortest Path Z-Scores of input position
}
\details{
This function calculates Shortest Path lengths of all nodes in the network via \code{shortest.paths} function
of igraph package. Sum shortest path lengts of every nodes, calculates and returns the Z-Scores of the input positions.
}
