% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortest_path_score.R
\name{shorteset_path_score}
\alias{shorteset_path_score}
\title{Shortest Path Score}
\usage{
shorteset_path_score(edge_list, filtered_info_df)
}
\arguments{
\item{edge_list}{list contains separate edge data.frames for each chain}

\item{filtered_info_df}{input data.frame which contain only one PDB entries}
}
\value{
Shortest Path Z-Scores of input position
}
\description{
Calculate Shortest Path Z-Scores of input position
}
\details{
This function calculates Shortest Path lengths of all nodes in the network via \code{shortest.paths} function
of igraph package. Sum shortest path lengts of every nodes, calculates and returns the Z-Scores of the input positions.
}
